+++ src/lib/services/validateForm.ts
@@ -1,10 +1,6 @@
 import { type DomainEvent } from '$lib/stores/ObjectStore';
 import { notificationStore } from '$lib/stores/notifications';
 
-interface FormError {
-  message: string;
-}
-
 export interface FormData {
   target: Record<string, { name: string; required: boolean; value: string }>;
 }
@@ -22,17 +18,15 @@ export function validateEmptyInput(formData: FormData): string[] {
   return errObject;
 }
 
-const buildTimeStamp = (date: string, time: string): Date => {
+export const buildTimeStamp = (date: string, time: string): Date => {
   return new Date(`${date}T${time}Z`);
 };
 
-export function validateEventData(event: DomainEvent): FormError[] {
-  //const event = get(EventStore);
-  const formErrors: FormError[] = [];
-
+export function validateEventData(event: DomainEvent): boolean {
   if (!event) {
     throw new Error('Event could be found');
   }
+
   const {
     startdate,
     starttime,
@@ -51,35 +45,55 @@ export function validateEventData(event: DomainEvent): FormError[] {
   const unpublishDateTime =
     unpublishdate && unpublishtime && buildTimeStamp(unpublishdate, unpublishtime).getTime();
 
+  let hasError = false;
+
   if (!startdate) {
     notificationStore.addToast('error', 'Mandatory start date is empty');
+    hasError = true;
   } else if (!starttime) {
     notificationStore.addToast('error', 'Mandatory start time is empty');
+    hasError = true;
+  }
+
+  if (!enddate) {
+    notificationStore.addToast('error', 'Mandatory end date is empty');
+    hasError = true;
+  } else if (!endtime) {
+    notificationStore.addToast('error', 'Mandatory end time is empty');
+    hasError = true;
   }
 
   if (new Date(startdate) < new Date()) {
     notificationStore.addToast('error', 'The start date cannot be in the past');
+    hasError = true;
   }
 
   if (startDateTime && endDateTime) {
     if (endDateTime < startDateTime) {
       notificationStore.addToast('error', 'The event end cannot be before the event start');
+      hasError = true;
     } else if (endDateTime === startDateTime) {
       notificationStore.addToast('warning', 'Both start and end date-time are the same');
+      hasError = true;
     }
   }
 
   if (startDateTime && publishDateTime && startDateTime < publishDateTime) {
-    notificationStore.addToast('error', 'The publish date cannot be after the start date');
+    notificationStore.addToast('error', 'The publish date cannot be later than the start date');
+    hasError = true;
   }
 
   if (unpublishDateTime && publishDateTime) {
     if (unpublishDateTime < publishDateTime) {
-      notificationStore.addToast('error', 'The unpublish date cannot be before the publish date');
+      notificationStore.addToast(
+        'error',
+        'The unpublish date cannot be earlier than the publish date'
+      );
+      hasError = true;
     } else if (unpublishDateTime === publishDateTime) {
       notificationStore.addToast('warning', 'Both publish and unpublish date-time are the same.');
+      hasError = true;
     }
-
-    return formErrors;
   }
+  return hasError;
 }